renamed to signed_delegate_action
