+++ src/commands/transaction/send_meta_transaction/mod.rs
@@ -1,14 +1,11 @@
-// mod network_for_transaction;
 mod sign_as;
-// mod transaction_signature_options;
 
 #[derive(Debug, Clone, interactive_clap::InteractiveClap)]
 #[interactive_clap(input_context = crate::GlobalContext)]
 #[interactive_clap(output_context = SendMetaTransactionContext)]
 pub struct SendMetaTransaction {
     /// Enter a signed delegate action as base64-encoded string
-    signed_delegate_action:
-        crate::types::signed_delegate_action_as_base64::SignedDelegateActionAsBase64,
+    signed_delegate_action: crate::types::signed_delegate_action::SignedDelegateActionAsBase64,
     #[interactive_clap(named_arg)]
     /// What is the relayer account ID?
     sign_as: self::sign_as::RelayerAccountId,
